test: merge geo test suites into 1
